// MiniCRMClient/src/api/rest/api/model/offer-file-datum-dto.ts
/**
 * CRM API
 * CRM API Reference
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { OfferFileType } from './offer-file-type';


export interface OfferFileDatumDto { 
    id?: number;
    created?: string;
    changed?: string;
    path?: string | null;
    type?: OfferFileType;
}