// MiniCRMClient/src/api/rest/api/model/offer-newsbreak-dto.ts
/**
 * CRM API
 * CRM API Reference
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserDto } from './user-dto';


export interface OfferNewsbreakDto { 
    id?: number;
    created?: string;
    changed?: string;
    text?: string | null;
    author?: UserDto;
}